refactor: #653 interview/page を Phase1 分割（types/Selection/Consent）
